Autorespond to people looking for prefix changes in the support server

+++ modules/support.js
@@ -4,14 +4,19 @@
  * ╚════ ║   (ocelotbotv5) support
  *  ════╝
  */
+
+const changePrefix = /.*(change|custom).*prefix.*/gi;
 module.exports = {
     name: "Support Server Specific Functions",
     init: function(bot){
         bot.client.on("message", bot.raven.wrap(async function onMessage(message) {
-            if(message.guild && message.guild.id === "322032568558026753"){
-                if(!message.author.bot && message.content.indexOf("discord.gg") > -1){
-                    message.delete();
-                }
+            if(message.guild && message.guild.id === "322032568558026753" && !message.author.bot){
+                if(message.content.indexOf("discord.gg") > -1)
+                    return message.delete();
+
+                if(changePrefix.exec(message.content))
+                    message.reply("To change the prefix, type !settings set prefix %\nWhere % is the prefix you want.");
+
             }
         }));
